Moving RDT 1.0 files to there own directory.
